AutoTender_Sovereign/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autotender"
version = "0.1.0"
description = "Automated document annotation and tender generator (educational use only)"
authors = [{ name="Your Name", email="you@example.com" }]
license = { text = "MIT" }
dependencies = [
    "pillow",
    "reportlab",
    "pytesseract",
    "pypdf"
]

[project.scripts]
autotender = "autotender.cli:main"

[tool.setuptools]
packages = ["autotender"]